+++ CHANGELOG.md
@@ -34,6 +34,7 @@
   - intercept and profiling hooks
   - on-demand convert standard model on-demand 
 - **Other**
+  - Video Preview: TAESD support for MiniMax
   - new optional transformer hooks: *settings -> compute add-ons*  
     *PAG: Perturbed attention guidance, PAB: Pyramid attention broadcast, FBC: First Block Cache, FC: Faster Cache, LS: Layer Skip, MC: Mag Cache, TS: TaylorSeer*  
     *note*: compatibility of different methods varies across different models